+++ CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.7.6 - 2022-12-27
+
+### Added
+* Store updates into temp data storage
+* Cache api data about operations
+
+
 ## 0.7.5 - 2022-12-26
 
 ### Added